moar precision

+++ cpp/pianolizer.hpp
@@ -7,14 +7,14 @@ class RingBuffer {
   private:
     unsigned mask;
     unsigned index = 0;
-    float* buffer;
+    double* buffer;
 
   public:
     RingBuffer(unsigned requestedSize) {
       const unsigned bits = ceil(log2(requestedSize + 1));
       const unsigned size = 1 << bits;
       mask = size - 1;
-      buffer = new float[size];
+      buffer = new double[size];
       for (unsigned i = 0; i < size; i++)
         buffer[i] = 0.;
     }
@@ -23,25 +23,25 @@ class RingBuffer {
       delete [] buffer;
     }
 
-    void write(float value) {
+    void write(double value) {
       index &= mask;
       buffer[index++] = value;
     }
 
-    float read(unsigned position) {
+    double read(unsigned position) {
       return buffer[(index + (~position)) & mask];
     }
 };
 
 class DFTBin {
   private:
-    float k, N;
-    float totalPower = 0.;
-    std::complex<float> coeff;
-    std::complex<float> dft = std::complex<float>(0., 0.);
+    double k, N;
+    double totalPower = 0.;
+    std::complex<double> coeff;
+    std::complex<double> dft = std::complex<double>(0., 0.);
 
   public:
-    float referenceAmplitude = 1.; // 0 dB level
+    double referenceAmplitude = 1.; // 0 dB level
 
     DFTBin(unsigned k_, unsigned N_)
       : k(k_), N(N_) {
@@ -50,31 +50,31 @@ class DFTBin {
       else if (N == 0)
         throw std::invalid_argument("N=0 is soooo not supported (Y THO?)");
 
-      coeff = exp(std::complex<float>(0., 2. * M_PI * (k / N)));
+      coeff = exp(std::complex<double>(0., 2. * M_PI * (k / N)));
     }
 
-    void update(float previousSample, float currentSample) {
+    void update(double previousSample, double currentSample) {
       totalPower += currentSample * currentSample;
       totalPower -= previousSample * previousSample;
 
-      dft = coeff * ((dft - std::complex<float>(previousSample, 0.)) + std::complex<float>(currentSample, 0.));
+      dft = coeff * ((dft - std::complex<double>(previousSample, 0.)) + std::complex<double>(currentSample, 0.));
     }
 
-    float rms() {
+    double rms() {
       return sqrt(totalPower / N);
     }
 
-    float amplitudeSpectrum() {
+    double amplitudeSpectrum() {
       return M_SQRT2 * (sqrt(norm(dft)) / N);
     }
 
-    float normalizedAmplitudeSpectrum() {
+    double normalizedAmplitudeSpectrum() {
       return totalPower > 0.
         ? amplitudeSpectrum() / rms()
         : 0.;
     }
 
-    float logarithmicUnitDecibels() {
+    double logarithmicUnitDecibels() {
       return 20. * log10(amplitudeSpectrum() / referenceAmplitude);
     }
 };
@@ -84,22 +84,22 @@ class MovingAverage {
     unsigned channels, sampleRate;
     int averageWindow = -1;
     unsigned targetAverageWindow;
-    float *sum;
+    double *sum;
 
     MovingAverage(unsigned channels_, unsigned sampleRate_)
       : channels(channels_), sampleRate(sampleRate_) {
-      sum = new float[channels];
+      sum = new double[channels];
     }
 
     ~MovingAverage() {
       delete [] sum;
     }
 
-    float averageWindowInSeconds() {
+    double averageWindowInSeconds() {
       return averageWindow / sampleRate;
     }
 
-    void averageWindowInSeconds(float value) {
+    void averageWindowInSeconds(double value) {
       targetAverageWindow = round(value * sampleRate);
       if (averageWindow == -1)
         averageWindow = targetAverageWindow;
@@ -112,7 +112,7 @@ class MovingAverage {
         averageWindow--;
     }
 
-    float read(unsigned n) {
+    double read(unsigned n) {
       return sum[n] / averageWindow;
     }
 };
@@ -123,10 +123,10 @@ class FastMovingAverage : public MovingAverage {
       : MovingAverage{ channels_, sampleRate_ }
     {}
 
-    void update(float levels[]) {
+    void update(double levels[]) {
       updateAverageWindow();
       for (unsigned n = 0; n < channels; n++) {
-        const float currentSum = sum[n];
+        const double currentSum = sum[n];
         sum[n] = averageWindow
           ? currentSum + levels[n] - currentSum / averageWindow
           : levels[n];
@@ -152,9 +152,9 @@ class HeavyMovingAverage : public MovingAverage {
         delete history[n];
     }
 
-    void update(float levels[]) {
+    void update(double levels[]) {
       for (unsigned n = 0; n < channels; n++) {
-        const float value = levels[n];
+        const double value = levels[n];
         history[n]->write(value);
         sum[n] += value;
